Big patch full of performance improvements This particularly benefits applications with a lot of nodes.

+++ panda/src/putil/simpleHashMap.h
@@ -42,12 +42,12 @@ public:
   INLINE Value &operator [] (const Key &key);
 
   INLINE size_t get_size() const;
-  INLINE bool has_element(int n) const;
-  INLINE const Key &get_key(int n) const;
-  INLINE const Value &get_data(int n) const;
-  INLINE Value &modify_data(int n);
-  INLINE void set_data(int n, const Value &data);
-  void remove_element(int n);
+  INLINE bool has_element(size_t n) const;
+  INLINE const Key &get_key(size_t n) const;
+  INLINE const Value &get_data(size_t n) const;
+  INLINE Value &modify_data(size_t n);
+  INLINE void set_data(size_t n, const Value &data);
+  void remove_element(size_t n);
 
   INLINE size_t get_num_entries() const;
   INLINE bool is_empty() const;
@@ -61,9 +61,9 @@ private:
 
   INLINE size_t get_hash(const Key &key) const;
 
-  INLINE bool is_element(int n, const Key &key) const;
-  INLINE void store_new_element(int n, const Key &key, const Value &data);
-  INLINE void clear_element(int n);
+  INLINE bool is_element(size_t n, const Key &key) const;
+  INLINE void store_new_element(size_t n, const Key &key, const Value &data);
+  INLINE void clear_element(size_t n);
   INLINE unsigned char *get_exists_array() const;
 
   void new_table();
